Added Zooming And Timeline
There are some major changes in this version. First off, I added code to
zoom the map onto a selected path, using the path's centroid and a set
amount of scaling. This got a little tricky becuase clicking on the
state a second time centered on the district. But I worked around it.

Next big change is a timeline at the bottom left, which redraws district
lines depending on which year you select.

Finally, I added a credits line and some small improvements to the chart
title. Now it never displays anything.